ssh-bruteforce-ids/requirements-realtime.txt
pandas
xgboost
scikit-learn
joblib
requests
python-dotenv
